refactor: Move SafeDOMXPath
