+++ devTools/phpstan.neon
@@ -75,7 +75,7 @@ parameters:
             message: "#queryCount#"
             identifier: return.type
             count: 1
-            path: ../src/TestFramework/SafeDOMXPath.php
+            path: ../src/TestFramework/XML/SafeDOMXPath.php
 
         # We are explicitly testing the guards here
         -
