+++ src/TestFramework/PhpUnit/Config/XmlConfigurationManipulator.php
@@ -42,7 +42,7 @@ use const FILTER_VALIDATE_URL;
 use function filter_var;
 use function implode;
 use Infection\TestFramework\PhpUnit\Config\Path\PathReplacer;
-use Infection\TestFramework\SafeDOMXPath;
+use Infection\TestFramework\XML\SafeDOMXPath;
 use const LIBXML_ERR_ERROR;
 use const LIBXML_ERR_FATAL;
 use const LIBXML_ERR_WARNING;
