+++ src/TestFramework/PhpUnit/Config/XmlConfigurationVersionProvider.php
@@ -35,7 +35,7 @@ declare(strict_types=1);
 
 namespace Infection\TestFramework\PhpUnit\Config;
 
-use Infection\TestFramework\SafeDOMXPath;
+use Infection\TestFramework\XML\SafeDOMXPath;
 use function Safe\preg_match;
 
 /**
