+++ src/TestFramework/XML/SafeDOMXPath.php
@@ -33,7 +33,7 @@
 
 declare(strict_types=1);
 
-namespace Infection\TestFramework;
+namespace Infection\TestFramework\XML;
 
 use DOMAttr;
 use DOMDocument;
