+++ tests/phpunit/TestFramework/XML/SafeDOMXPath/invalid.xml
